+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Brewtils Changelog
 ------
 TBD
 
+- Added support for display name to command decorator
 - Updated Wait Timeout Exception expected HTTP code from 408 to 504
 - Dropping Official Python 2.7 Support